requirements.txt
alembic==1.8.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
chardet==4.0.0
cliff==4.0.0
cloudpickle==2.2.1
cmaes==0.8.2
colorlog==6.7.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.39.2
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.27.1
gymnasium-notices==0.0.1
idna==2.10
importlib-metadata==6.1.0
ipywidgets==8.0.4
jax-jumpy==1.0.0
Jinja2==3.1.2
jupyter==1.0.0
jupyter-console==6.4.4
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
mpmath==1.2.1
networkx==3.0
numpy==1.22.3
optuna==3.0.1
packaging==23.0
PettingZoo==1.18.1
Pillow==9.1.1
pyglet==1.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.25.1
scipy==1.10.1
six==1.16.0
sympy==1.11.1
tensorflow==2.3.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
typing_extensions==4.4.0
urllib3==1.26.13
wincertstore==0.2
zipp==3.15.0
